move GetKOSThreshold to bobtail/bobtail.h
